requirements.txt
click
requests
curl_cffi
beautifulsoup4
lxml         
rich
pillow
img2pdf